Double to int

+++ Algebra/src/main/python/ppcm.py
@@ -7,7 +7,7 @@ def ppcm(n1: int, n2: int) -> int:
         n1 = n2
         n2 = remainder
         remainder = n1 % n2
-    return product / n2
+    return int(product / n2)
 
 n1: int = 60
 n2: int = 36
